// apps/shop-corner-api/src/libs/enums/common.enum.ts
export enum Direction {
	ASC = 1,
	DESC = -1,
}